+++ src/SecucardConnect/Product/General/Model/CheckoutOptions.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace SecucardConnect\Product\General\Model;
+
+use SecucardConnect\Product\Common\Model\BaseModel;
+
+/**
+ * CheckoutOptions Api Model class
+ */
+class CheckoutOptions
+{
+    /**
+     * @var array
+     */
+    public $background;
+}
+